requirements.txt
alembic
fastapi
fastapi-sqlalchemy
psycopg2
pydantic
python-dotenv
sqlalchemy
uvicorn
python-jose
werkzeug